--- rosetta/converter/address.go
@@ -1,1 +1,0 @@
-package converter
