--- rosetta/converter/amount.go
@@ -1,1 +1,0 @@
-package converter
