--- rosetta/converter/block.go
@@ -1,1 +1,0 @@
-package converter
