+++ rosetta/converter/converter.go
@@ -1,5 +1,12 @@
 package converter
 
+import (
+	"strconv"
+
+	"github.com/awfm9/flow-dps/model/identifier"
+	"github.com/awfm9/flow-dps/model/rosetta"
+)
+
 type Converter struct {
 }
 
@@ -8,4 +15,12 @@ func New() *Converter {
 	c := &Converter{}
 
 	return c
+}
+
+func (c *Converter) Balance(currency identifier.Currency, balance uint64) rosetta.Amount {
+	amount := rosetta.Amount{
+		Value:    strconv.FormatUint(balance, 10),
+		Currency: currency,
+	}
+	return amount
 }
